+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 
 *   Generate file comments by default.
 *   When a comment begins with multiple percent signs, convert all of them to hashes.
+*   Separate clauses within a case, receive, or catch leaked variable scopes to each other. Fixed.
+*   Support remote function calls with an expression as the function name.
 
 ## v0.0.3 (2016-01-03)